--- packages/delegate/src/externalObjects.ts
@@ -1,29 +1,0 @@
-import { GraphQLSchema, GraphQLError } from 'graphql';
-
-import { SubschemaConfig, ExternalObject } from './types';
-import { OBJECT_SUBSCHEMA_SYMBOL, FIELD_SUBSCHEMA_MAP_SYMBOL, UNPATHED_ERRORS_SYMBOL } from './symbols';
-
-export function isExternalObject(data: any): data is ExternalObject {
-  return data[UNPATHED_ERRORS_SYMBOL] !== undefined;
-}
-
-export function annotateExternalObject(
-  object: any,
-  errors: Array<GraphQLError>,
-  subschema: GraphQLSchema | SubschemaConfig | undefined
-): ExternalObject {
-  Object.defineProperties(object, {
-    [OBJECT_SUBSCHEMA_SYMBOL]: { value: subschema },
-    [FIELD_SUBSCHEMA_MAP_SYMBOL]: { value: Object.create(null) },
-    [UNPATHED_ERRORS_SYMBOL]: { value: errors },
-  });
-  return object;
-}
-
-export function getSubschema(object: ExternalObject, responseKey: string): GraphQLSchema | SubschemaConfig {
-  return object[FIELD_SUBSCHEMA_MAP_SYMBOL][responseKey] ?? object[OBJECT_SUBSCHEMA_SYMBOL];
-}
-
-export function getUnpathedErrors(object: ExternalObject): Array<GraphQLError> {
-  return object[UNPATHED_ERRORS_SYMBOL];
-}
