+++ packages/delegate/src/index.ts
@@ -4,7 +4,6 @@ export * from './applySchemaTransforms';
 export * from './createRequest';
 export * from './defaultMergedResolver';
 export * from './delegateToSchema';
-export * from './externalObjects';
 export * from './resolveExternalValue';
 export * from './subschemaConfig';
 export * from './types';
